Feature: Party doesn't start automatically anymore
Party doesn’t start automatically anymore when admin confirms email. Instead party is in a frozen state, no emails are sent and all changes are possible. When the admin decides to start the party, the participants are shuffled and the mails are sent out. When started with excludes, limited changes are possible. When started without, most changes are possible, but not encouraged. Workaround for #151
