+++ src/Intracto/SecretSantaBundle/Resources/translations/messages.de.yml
@@ -80,18 +80,19 @@ pool-create:
         participants: Teilnehmer
 
 # Pool/exclude.html.twig
-pool-exclude:
+party_manage-exclude:
     title: Ausschlieβen
     description: Mit dieser Liste können Sie auch Kombinationen von Teilnehmern ausschließen. o können Sie zum Beispiel verhindern, dass Mitglieder derselben Familie Geschenke füreinander kaufen solle.
-
+#    information
     placeholder_exclude: Um den Teilnehmer auszuwählen, der ausgeschlossen werden soll, klicken Sie auf dessen Namen.
-
     btn:
-        create_event: Erstellen Sie Ihr Ereignis!
-
+#        start_party:
+        cancel: Stornieren
     label:
         name: Name
         exclude: Ausschließen
+    feedback:
+#        not_enough:
 
 # Pool/created.html.twig
 pool-created:
@@ -309,6 +310,8 @@ pool_manage_valid:
 #        pool_update: ?
 #        remove_participant_confirm: ?
 #        updated_party: ?
+#        start_party:
+#        start_party_with_excludes:
 
     label:
         name: Name
@@ -464,6 +467,9 @@ flashes:
 #            success: ?
 #            danger: ?
 #        updated_party:
+#            success: ?
+#            danger: ?
+#        start_party:
 #            success: ?
 #            danger: ?
