+++ src/Intracto/SecretSantaBundle/Resources/translations/messages.pl.yml
@@ -70,18 +70,19 @@ pool-create:
         participants: Uczestników
 
 # Pool/exclude.html.twig
-pool-exclude:
+party_manage-exclude:
     title: Wyłącz
     description: Ta lista może potencjalnie wykluczyć pewne kombinacje uczestników. Na przykład uniemożliwić członkom tej samej rodziny wymianę prezentów między sobą.
-
+#    information:
     placeholder_exclude: Kliknij i wybierz uczestników, których chcesz wykluczyć
-
     btn:
-        create_event: Utwórz wydarzenie!
-
+#        start_party:
+#        cancel:
     label:
         name: Imię
         exclude: Wyłącz
+    feedback:
+#       not_enough:
 
 # Pool/created.html.twig
 pool-created:
@@ -271,6 +272,8 @@ pool_manage_valid:
 #        pool_update: ?
 #        remove_participant_confirm: ?
 #        updated_party: ?
+#        start_party:
+#        start_party_with_excludes:
 
     label:
         name: Imię
@@ -422,6 +425,9 @@ flashes:
 #            success: ?
 #            danger: ?
 #        updated_party:
+#            success: ?
+#            danger: ?
+#        start_party:
 #            success: ?
 #            danger: ?
