+++ src/Intracto/SecretSantaBundle/Resources/translations/messages.ru.yml
@@ -77,18 +77,19 @@ pool-create:
         participants: участников
 
 # Pool/exclude.html.twig
-pool-exclude:
+party_manage-exclude:
     title: Исключить:
     description: Список позволит вам на ваш выбор предотвратить нежелательные комбинации получателей и дарителей. Например, это поможет членам одной семьи не выбирать подарки друг для друга.
-
+#    information:
     placeholder_exclude: Кликните для выбора участников, которых вы хотели бы исключить
-
     btn:
-        create_event: Создать событие!
-
+#        start_party:
+#        cancel:
     label:
         name: Имя
         exclude: Исключить
+    feedback:
+#       not_enough:
 
 # Pool/created.html.twig
 pool-created:
@@ -309,6 +310,8 @@ pool_manage_valid:
 #        pool_update: ?
 #        remove_participant_confirm: ?
 #        updated_party: ?
+#        start_party:
+#        start_party_with_excludes:
 
     label:
         name: Имя
@@ -464,6 +467,9 @@ flashes:
 #            success: ?
 #            danger: ?
 #        updated_party:
+#            success: ?
+#            danger: ?
+#        start_party:
 #            success: ?
 #            danger: ?
